Revert "Added support for jina-reranker-v1-turbo-en and bring-your-own reranking models"
This reverts commit e32f7a9362d1c1b0e7ebba881b0fda00e7400beb.
