+++ src/lib.rs
@@ -56,10 +56,10 @@ mod tests;
 
 pub use ort::ExecutionProviderDispatch;
 
-pub use crate::common::{read_file_to_bytes, Embedding, TokenizerFiles, UserDefinedModel};
+pub use crate::common::{read_file_to_bytes, Embedding};
 pub use crate::models::reranking::{RerankerModel, RerankerModelInfo};
 pub use crate::models::text_embedding::{EmbeddingModel, ModelInfo};
 pub use crate::reranking::{RerankInitOptions, RerankResult, TextRerank};
 pub use crate::text_embedding::{
-    InitOptions, InitOptionsUserDefined, TextEmbedding
+    InitOptions, InitOptionsUserDefined, TextEmbedding, TokenizerFiles, UserDefinedEmbeddingModel,
 };
