+++ src/models/reranking.rs
@@ -2,25 +2,15 @@
 pub enum RerankerModel {
     /// BAAI/bge-reranker-base
     BGERerankerBase,
-    // jinaai/jina-reranker-v1-turbo-en
-    JINARerankerV1TurboEn,
 }
 
 pub fn reranker_model_list() -> Vec<RerankerModelInfo> {
-    let reranker_model_list = vec![
-        RerankerModelInfo {
-            model: RerankerModel::BGERerankerBase,
-            description: String::from("reranker model for English and Chinese"),
-            model_code: String::from("BAAI/bge-reranker-base"),
-            model_file: String::from("onnx/model.onnx"),
-        },
-        RerankerModelInfo {
-            model: RerankerModel::JINARerankerV1TurboEn,
-            description: String::from("reranker model for English"),
-            model_code: String::from("jinaai/jina-reranker-v1-turbo-en"),
-            model_file: String::from("onnx/model.onnx"),
-        },
-    ];
+    let reranker_model_list = vec![RerankerModelInfo {
+        model: RerankerModel::BGERerankerBase,
+        description: String::from("reranker model for english and chinese"),
+        model_code: String::from("BAAI/bge-reranker-base"),
+        model_file: String::from("onnx/model.onnx"),
+    }];
     reranker_model_list
 }
